Implements cross-service data fetching
Introduces an aspect-oriented approach to fetch and enrich data from other services.

This change introduces a `CrossServiceConfig` aspect that intercepts controller responses and uses annotations to define how data from other services should be fetched and merged into the response. This eliminates the need for repetitive code in each controller and centralizes the logic for cross-service data integration.

The gateway and service dependencies are updated to the latest versions.

A new project entity and controller are added to service-1, to use the newly created cross-service config.
The new user controller and service are created in service-2.